% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_ca_certificate}
\alias{iot_describe_ca_certificate}
\title{Describes a registered CA certificate}
\usage{
iot_describe_ca_certificate(certificateId)
}
\arguments{
\item{certificateId}{[required] The CA certificate identifier.}
}
\description{
Describes a registered CA certificate.
}
\section{Request syntax}{
\preformatted{iot$describe_ca_certificate(
  certificateId = "string"
)
}
}

\keyword{internal}
