% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_role_alias}
\alias{iot_describe_role_alias}
\title{Describes a role alias}
\usage{
iot_describe_role_alias(roleAlias)
}
\arguments{
\item{roleAlias}{[required] The role alias to describe.}
}
\description{
Describes a role alias.
}
\section{Request syntax}{
\preformatted{iot$describe_role_alias(
  roleAlias = "string"
)
}
}

\keyword{internal}
