% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_stream}
\alias{iot_describe_stream}
\title{Gets information about a stream}
\usage{
iot_describe_stream(streamId)
}
\arguments{
\item{streamId}{[required] The stream ID.}
}
\description{
Gets information about a stream.
}
\section{Request syntax}{
\preformatted{iot$describe_stream(
  streamId = "string"
)
}
}

\keyword{internal}
