% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing_group}
\alias{iot_describe_thing_group}
\title{Describe a thing group}
\usage{
iot_describe_thing_group(thingGroupName)
}
\arguments{
\item{thingGroupName}{[required] The name of the thing group.}
}
\description{
Describe a thing group.
}
\section{Request syntax}{
\preformatted{iot$describe_thing_group(
  thingGroupName = "string"
)
}
}

\keyword{internal}
