% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_enable_topic_rule}
\alias{iot_enable_topic_rule}
\title{Enables the rule}
\usage{
iot_enable_topic_rule(ruleName)
}
\arguments{
\item{ruleName}{[required] The name of the topic rule to enable.}
}
\description{
Enables the rule.
}
\section{Request syntax}{
\preformatted{iot$enable_topic_rule(
  ruleName = "string"
)
}
}

\keyword{internal}
