% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_ota_update}
\alias{iot_get_ota_update}
\title{Gets an OTA update}
\usage{
iot_get_ota_update(otaUpdateId)
}
\arguments{
\item{otaUpdateId}{[required] The OTA update ID.}
}
\description{
Gets an OTA update.
}
\section{Request syntax}{
\preformatted{iot$get_ota_update(
  otaUpdateId = "string"
)
}
}

\keyword{internal}
