% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_scheduled_audit}
\alias{iot_update_scheduled_audit}
\title{Updates a scheduled audit, including what checks are performed and how
often the audit takes place}
\usage{
iot_update_scheduled_audit(frequency, dayOfMonth, dayOfWeek,
  targetCheckNames, scheduledAuditName)
}
\arguments{
\item{frequency}{How often the scheduled audit takes place. Can be one of "DAILY",
"WEEKLY", "BIWEEKLY" or "MONTHLY". The actual start time of each
audit is determined by the system.}

\item{dayOfMonth}{The day of the month on which the scheduled audit takes place. Can be
"1" through "31" or "LAST". This field is required if the
"frequency" parameter is set to "MONTHLY". If days 29-31 are
specified, and the month does not have that many days, the audit takes
place on the "LAST" day of the month.}

\item{dayOfWeek}{The day of the week on which the scheduled audit takes place. Can be one
of "SUN", "MON", "TUE", "WED", "THU", "FRI" or "SAT". This
field is required if the "frequency" parameter is set to "WEEKLY" or
"BIWEEKLY".}

\item{targetCheckNames}{Which checks are performed during the scheduled audit. Checks must be
enabled for your account. (Use \code{DescribeAccountAuditConfiguration} to
see the list of all checks including those that are enabled or
\code{UpdateAccountAuditConfiguration} to select which checks are enabled.)}

\item{scheduledAuditName}{[required] The name of the scheduled audit. (Max. 128 chars)}
}
\description{
Updates a scheduled audit, including what checks are performed and how
often the audit takes place.
}
\section{Request syntax}{
\preformatted{iot$update_scheduled_audit(
  frequency = "DAILY"|"WEEKLY"|"BIWEEKLY"|"MONTHLY",
  dayOfMonth = "string",
  dayOfWeek = "SUN"|"MON"|"TUE"|"WED"|"THU"|"FRI"|"SAT",
  targetCheckNames = list(
    "string"
  ),
  scheduledAuditName = "string"
)
}
}

\keyword{internal}
