% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_validate_security_profile_behaviors}
\alias{iot_validate_security_profile_behaviors}
\title{Validates a Device Defender security profile behaviors specification}
\usage{
iot_validate_security_profile_behaviors(behaviors)
}
\arguments{
\item{behaviors}{[required] Specifies the behaviors that, when violated by a device (thing), cause
an alert.}
}
\description{
Validates a Device Defender security profile behaviors specification.
}
\section{Request syntax}{
\preformatted{iot$validate_security_profile_behaviors(
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          )
        ),
        durationSeconds = 123,
        consecutiveDatapointsToAlarm = 123,
        consecutiveDatapointsToClear = 123,
        statisticalThreshold = list(
          statistic = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
